Replace jshasher with native Web Crypto API and md5 with sha1
We can get rid of another dependency by replacing the jshasher package
with native functionality from the Web Crypto API.
One caveat: MD5 hashes are not implemented here, so we need to use a
different hash. For that reason I've choosen SHA-1, as I think it might
be a bit faster than the longer SHA hashes.